More code style fixes

+++ lcDXFDWG/libdxfrw/dxfimpl.cpp
@@ -186,7 +186,7 @@ void DXFimpl::addLayer(const DRW_Layer& data) {
     }
 
     auto lp = _document->linePatternByName(data.lineType);
-    auto isFrozen = (bool) (data.flags & 1u);
+    auto isFrozen = (bool) ((unsigned int) data.flags & 1u);
 
     auto layer = std::make_shared<lc::Layer>(data.name, lw->width(), col->color(), lp, isFrozen);
     // If a layer starts with a * it's a special layer we don't process yet
@@ -412,7 +412,7 @@ void DXFimpl::addLWPolyline(const DRW_LWPolyline& data) {
         points.emplace_back(lc::geo::Coordinate(i->x, i->y), i->bulge, i->stawidth, i->endwidth);
     }
 
-    auto isCLosed = data.flags & 0x01u;
+    auto isCLosed = (unsigned int) data.flags & 0x01u;
     auto lcLWPolyline = std::make_shared<lc::entity::LWPolyline>(
             points,
             data.width,
